Move tests to the right file.
